Layout: wrap DesignPreview in SitePreview
SitePreview is a generic wrapper for components that display a preview (a `WebPreview` component) as an overlay. It is intended to be a singleton; only one `SitePreview` should ever be rendered. What it renders as its children is controlled by its props.

WebPreview's many props and support for children mean that it can be quite different depending on its context (eg: simple site previews, page and post previews, editor previews, and customizer previews all use different props and children).

Since there are many different ways that `WebPreview` can be used, this component acts as a container to decide which components appear. The component choice is controlled via a prop called `currentPreviewType` which can be set via the Redux store. That key is used to determine what is rendered by SitePreview.

Right now `currentPreviewType` defaults to 'design-preview' so that the existing behavior is unchanged, but it is easy to add different preview components.